+++ scripts/database/otrs-schema.mssql.sql
@@ -643,7 +643,7 @@ CREATE INDEX index_time_accounting_ticket_id ON time_accounting (ticket_id);
 -- ----------------------------------------------------------
 CREATE TABLE ticket_watcher (
     ticket_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    user_id INTEGER NOT NULL,
     create_time DATETIME NOT NULL,
     create_by INTEGER NOT NULL,
     change_time DATETIME NOT NULL,
